[move-compiler v2] Comparison testing for file-format generator (#9345)
* [move-compiler v2] File-format generator first version

This adds an initial version of the generation of the external representation of Move code, the so-called file format.

- Compiles from the register machine of the compiler IR into the stack machine the VM uses. Some attempts are made to optimize usage of the stack, though more can/must be done to optimize the stack machine code.
- Adds a set of basic tests, but more e2e execution tests should be added. (Subsequent PRs.)
- The translation can benefit from analysis results like live-var which allows it make smarter use of `CopyLoc` vs `MoveLoc` (to be added in the future)

* Addressing reviewer comments

* More review fixes

* [move-compiler v2] Comparison testing for file-format generator

This extends the transactional-test-runner to support compiler V2 _and_ to do comparison testing between V1 and V2.

It ports over most of the transactional tests of the V1 compiler into the V2 tree and runs comparison testing on them, fixing bugs throughout the compilation chain as needed.

Closes #9334

GitOrigin-RevId: 53913b462ccc838c787a264ee3ff9a1367484b7b