[compiler-v2] Adding logging to the compiler (#11840)
* [compiler-v2] Adding logging to the compiler

Until now we did not had a systematic way how to log information for debugging purposes. This uses the popular [`log`](https://crates.io/crates/log) facade as a basic API for logging. The macros `error!`, `warn!`, `info!`, `debug!`, and `trace!` can be used like `println!`.  The backend we are installing for the compiler is configured via an an env var and described in the header of the new `logging.rs` module. See there for more discussion.

A few first log instructions are also inserted in the code.

* Addressing reviewer comments

* Making log level determinator for debugging. Dumping bytecode to log instead of file. Removing feature verbose-debug-print and express it programmatically instead.

GitOrigin-RevId: 19704a3a0d4e4ad0941125e6a1c9256f7abfdc38