[compiler-v2] Turns on critical edge elimination since it is needed by enums (#14993)
* [compiler-v2] Turns on critical edge elimination since it is needed by enums

Enum matches of following form create critical edges in the control flow graph which we haven't seen before:

```
match (self) {
   Person{id} if *id > 0 => *id,
   Institution{id, ..} => *id,
   _ => 0
}
```

That is because there are two incoming edges into the `Institution` match branch. The first is if the `Person` pattern doesn't match, the 2nd is if the pattern condition is false.

Critical edge elimination was turned off before. This PR turns it on, closing #14733.

* Refactor test, moving it over to transactional testsuite to also test behavior, and adding more cases.

GitOrigin-RevId: c9315e394d460dd120dc472ba14a3d566772cf47